refactor interfaces out to get around testing restrictions
